# openapi.yaml
...
openapi: 3.0.3
info:
  title: APIFoundry
  version: "0.1.0"
  description: >
    General-purpose mock API server used as a testing target for functional, negative, automation,
    contract, and performance/load testing. This foundation slice provides configuration, health,
    version/info, and the shared error/pagination envelopes every other endpoint reuses.
servers:
  - url: /
...
paths:
...
  /cookies:
    get:
      tags: [HTTP Testing Utilities]
      summary: Reports every cookie present on the request
      operationId: getCookies
      security: []
      responses:
        "200":
          description: Current cookies
          headers:
            X-Request-ID: { $ref: "#/components/headers/X-Request-ID" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CookieMapResponse" }
    post:
      tags: [HTTP Testing Utilities]
      summary: Sets an arbitrary named cookie without disturbing any other cookie held
      operationId: postCookies
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CookieSetRequest" }
      responses:
        "200":
          description: Cookie set
          headers:
            X-Request-ID: { $ref: "#/components/headers/X-Request-ID" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CookieSetResponse" }
        "400": { $ref: "#/components/responses/ValidationError" }
    delete:
      tags: [HTTP Testing Utilities]
      summary: Clears one named cookie via ?name=, leaving every other cookie intact
      operationId: deleteCookies
      security: []
      parameters:
        - name: name
          in: query
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        "200":
          description: Cookie cleared (idempotent, even if it was never set)
          headers:
            X-Request-ID: { $ref: "#/components/headers/X-Request-ID" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CookieClearResponse" }
        "400": { $ref: "#/components/responses/ValidationError" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    basicAuth:
      type: http
      scheme: basic
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details, requestId]
          properties:
            code:
              type: string
              example: RESOURCE_NOT_FOUND
            message:
              type: string
            details:
              type: object
            requestId:
              type: string
              format: uuid
...
    CookieMapResponse:
      type: object
      required: [cookies]
      properties:
        cookies:
          type: object
          additionalProperties: { type: string }
    CookieSetRequest:
      type: object
      required: [name, value]
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1 }
        value: { type: string }
    CookieSetResponse:
      type: object
      required: [name, value]
      properties:
        name: { type: string }
        value: { type: string }
    CookieClearResponse:
      type: object
      required: [cleared, name]
      properties:
        cleared: { type: boolean, example: true }
        name: { type: string }
  responses:
...
    ValidationError:
      description: The request failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  headers:
    X-Request-ID:
      description: Unique identifier for this request, echoed on every response and in every error body.
      schema:
        type: string
        format: uuid